Modo reunião: respostas da Alma entram na ata + só responde quando chamada (determinístico)
Duas correções pedidas pelo Rui (2026-07-31):

1. "quero que as coisas ditas pela alma e relevantes tb façam parte da
   acta": reuniao.registar_resposta_alma regista a resposta do Claude
   (via /alma/reuniao/pergunta_empresa) na transcrição acumulada, usada
   tanto para dar contexto a perguntas seguintes como para o resumo/ata
   final — antes só o que as pessoas diziam entrava nesse registo.

   Usa um índice fracionário (o índice mais recente conhecido do lado do
   cliente, menos 0.5) em vez de "o mais alto + 1": os índices dos turnos
   humanos são atribuídos pelo cliente de forma independente, por isso
   "+1" acabaria por colidir com um índice que o cliente ainda ia usar,
   sobrescrevendo a resposta da Alma silenciosamente (apanhado em teste
   antes de publicar). db.carregar_estado_reuniao passa a fazer float(),
   não int(), ao reidratar o estado persistido, para o índice fracionário
   sobreviver a um reinício do servidor a meio de uma reunião.

2. "a alma esta a intervir demasiadas vezes mesmo que o seu nome nao seja
   chamado": a sessão de conversação tinha create_response=true (omissão
   da API) — respondia sozinha a qualquer turno detetado, confiando só na
   instrução "só respondes quando chamada", que não chegou para garantir
   isso com fiabilidade. Muda para create_response=false: a sessão só
   transcreve e deteta turnos, nunca gera resposta sozinha; é o browser
   que decide, por uma verificação determinística (/\balma\b/i no texto
   transcrito), quando disparar response.create — voltando ao mesmo nível
   de garantia que a versão anterior (só transcrição) tinha por código,
   em vez de depender do juízo do modelo. interrupt_response mantém-se
   true: uma fala nova continua a cortar uma resposta em curso.

+++ db.py
@@ -701,7 +701,12 @@ def carregar_estado_reuniao(sessao: str):
             linha = cur.fetchone()
     if not linha:
         return None
-    excertos = {int(indice): texto for indice, texto in linha["excertos"].items()}
+    # float, não int — uma resposta da Alma regista-se num índice fracionário
+    # (ver tools/reuniao.py:registar_resposta_alma), para ficar sempre entre
+    # o turno que a desencadeou e o seguinte, sem colidir com índices
+    # inteiros futuros atribuídos pelo cliente (que não sabe desta inserção
+    # do lado do servidor)
+    excertos = {float(indice): texto for indice, texto in linha["excertos"].items()}
     return {"excertos": excertos, "processados": linha["processados"]}
 
 def eliminar_estado_reuniao(sessao: str):
